Fixes for using OFI shared memory provider
OneCCL has the capability to use two end points,a shared memory provider endpoint for intra node and a network provider endpoint for internode.

The use of shared memory provider is enabled by environment variable CCL_ATL_SHM. However two bugs in the code base does not allow the use of OFI shm provider.

The commit has the fixes for the two bug.